+++ config.js
@@ -2,7 +2,7 @@ module.exports = {
   //number of pages to write at a time, to the queue
   "batch_size": 1000,
   //the default name of the collection to write to
-  "collection": "wikipedia",
+  "collection": "pages",
   //update interval
   "logInterval": 10000,
 }