+++ scratch.js
@@ -1,19 +1,23 @@
 const dumpster = require('./src')
 const drop = require('./src/lib/drop-db')
 
-const path = '/Users/spencer/data/wikipedia/afwiki-latest-pages-articles.xml' //4.3mins
-// const path = '/Users/spencer/data/wikipedia/simplewiki-latest-pages-articles.xml'
-// const path = '/Users/spencer/data/wikipedia/eswiki-latest-pages-articles.xml'
-// const path = '/Users/spencer/data/wikipedia/enwiki-latest-pages-articles.xml'
+// const path = '/Users/spencer/data/wikipedia/afwiki-latest-pages-articles.xml' //4.3mins
+// const path = '/Users/spencer/data/wikipedia/simplewiki-latest-pages-articles.xml' //5mins
+const path = '/Users/spencer/data/wikipedia/eswiki-latest-pages-articles.xml' //2hrs
+// const path = '/Users/spencer/data/wikipedia/enwiki-latest-pages-articles.xml'  //6hrs
 // const path = './tests/smallwiki-latest-pages-articles.xml'
 // const path = './tests/tinywiki-latest-pages-articles.xml'
 const dbName = path.match(/\/([a-z-]+)-latest-pages/)[1]
 let options = {
   file: path,
   db: dbName,
-// plaintext: false,
-// html: false,
-// markdown: false,
+  custom: function(doc) {
+    return {
+      _id: doc.title(),
+      title: doc.title(),
+      categories: doc.categories(),
+    }
+  }
 }
 
 //delete all pages