+++ src/worker/02-parseWiki.js
@@ -24,7 +24,12 @@ const parseData = function(page, options) {
       return null
     }
     //turn the wtf_wikipedia document into storable json
-    let data = doc.json(options)
+    let data = {}
+    if (!options.custom) { //default format
+      data = doc.json(options)
+    } else { //DIY format
+      data = options.custom(doc)
+    }
     data.title = page.title || data.title
     data = encode.encodeData(data);
     //use the title/pageID from the xml