+++ changelog.md
@@ -16,3 +16,7 @@
 * rename from `wikipedia-to-mongo` to `dumpster-dive`
 * use wtf_wikipedia v3 (a big re-factor too!)
 * use `line-by-line`, and `worker-nodes` to run parsing in parallel
+### v3.1.0
+* fix connection time-outs & improve logging output
+* change default collection name to `pages`
+* add `.custom()` function support